Refactor yield from expression rewriting
