+++ src/transform/tests_rewrite_expr_to_stmt.txt
@@ -80,3 +80,35 @@ def _dp_gen_1(items):
         else:
             yield i
 x = _dp_gen_1(__dp__.iter(items))
+
+$ rewrites yield from expression
+
+x = yield from y
+=
+_dp_yield_from_state_2 = __dp__.yield_from_init(y)
+_dp_yield_from_sent_3 = None
+while True:
+    _dp_tmp_4 = __dp__.ne(__dp__.getitem(_dp_yield_from_state_2, 0), __dp__.RUNNING)
+    if _dp_tmp_4:
+        break
+    try:
+        _dp_yield_from_sent_3 = yield __dp__.getitem(_dp_yield_from_state_2, 1)
+    except:
+        _dp_yield_from_state_2 = __dp__.yield_from_except(_dp_yield_from_state_2, __dp__.current_exception())
+    else:
+        _dp_yield_from_state_2 = __dp__.yield_from_next(_dp_yield_from_state_2, _dp_yield_from_sent_3)
+_dp_tmp_1 = __dp__.getitem(_dp_yield_from_state_2, 1)
+x = _dp_tmp_1
+
+$ rewrites named expression in boolop
+
+if (y := foo()) and bar:
+    pass
+=
+_dp_tmp_2 = foo()
+y = _dp_tmp_2
+_dp_tmp_1 = _dp_tmp_2
+if _dp_tmp_1:
+    _dp_tmp_1 = bar
+if _dp_tmp_1:
+    pass
